refactor(scancode): Move `toSummary()` to `ScanCodeResultModelMapper`
